Add (experimental) support for setting brightness on thaw/resume.
I can confirm that running `bri thaw` or `bri resume` will do the trick,
but from a quick test it doesn't seem like it's working on my laptop.
This may be because the computer isn't really "suspended" (I'm just closing it
and reopening it, only waiting about a minute). Maybe it needs to be closed
for longer for it to count as suspension and for this to work out right.

Additionally, I revised part of bri.sh, in an attempt to refactor it.