Smooth out one unnecessary bashism - change `source` to `.`
I used Ubuntu's `checkbashisms` program on bri.sh.
It identified two issues: using `source` instead of `.`
and one instance of `[[`. I changed the `source` one since it was simple to do
so.

However I was not sure how to get rid of the double-brackets. I'd like to make
this more portable by not using them, but I do not know how to change this
statement to using single brackets. Another day perhaps.