Add $bri_minimum setting, so that the minimum is configurable.
On my system, 0 is actually not a great minimum, and 1 is better.
This has to do with the fact that 100 divided by the ACPI max brightness
is not always an integer. On my system it's 14.2... So you have to pick an
interval that is close, then use the minimum value to offset it as needed.